requirements.txt
pandas>=1.5.0
numpy>=1.21.0
aiohttp>=3.8.0
requests==2.31.0
websockets==12.0
numpy==1.24.3
pandas==2.0.3
asyncio
PyYAML==6.0.1
aiofiles==24.1.0
python-dateutil==2.8.2
asyncio-mqtt>=0.11.0
TA-Lib>=0.4.25
